type_traits: improve SelectBySize error reporting.
Previously, if SelectBySize could not find a type of the given size
in its parameter list, it would hit a base case where the template
was not specialized -- producing the stereotypical cascade of template
error messages.  In GCC the most specific message was something like
'invalid use of incomplete type'.

Here, I've added a base case that terminates recursion in a static
assert with a more readable error message.

Notice that the base case still takes a single type parameter --
i.e. it does not completely exhaust the list before reporting.  This
is to ensure that the static_assert is conditional on something.
Otherwise, it will fire on every compile.