mem/arena: simplify Arena.
I've decided that m4vgalib's odd prioritized non-contiguous region
scheme is too specific to that application to justify the complexity
here, in the reusable bits.  Fortunately, refactoring that as a layer
atop a simpler Arena proved pretty easy.

Arena now manages a single region, making it simpler to implement but
also to create and use.