math/vector: generalize combinators, shuffle.
I noticed after the first draft that the only difference between each
specialization of the unary/binary/reduce combinators was the shape of
the vector...yet all vectors have similar shapes.

Turns out, all we really need to generalize the combinators is a way of
accessing elements by index.  So I've added a 'get' member template to
do just that, and pushed the templates down into the Vector subtemplate.

An overload of that template allows for arbitrary shuffle operations.
The syntax is not as nice as GLSL, but it's acceptable given the
language I'm working with.

While I was in there, I replaced the arithmetic lambdas with functors
from the new functors header, because C++11 is unimaginative at compile
time.