math/matrix: use real constructors.
'Matrix<r, c, T>' now has a constructor that accepts 'r' vectors of
length 'c'.  This allows literal matrices to be written without the
double-brace syntax we needed before, while still allowing for inference
so that the vector type name need not be written.

The approach I had to use is a little complicated, but seems solid, and
I'm inclined to apply it to Vector next.