math/vector: overload normalized for unit vectors.
I was debating this.  The fact of the matter is that unit vectors may
actually drift away from unit length, particularly as they are e.g.
rotated by quaternions.  So it might actually be useful to normalize
them explicitly.

But this doesn't take away your ability to do this.  You can always
assert that the vector may not be normal by casting back to a Vector.

Defining UnitVector-specific operations is key to having algorithms
magically become cheaper when used on units.