Replace ARMv7M saturation functions with general one.
GCC will *usually* replace this 'clamp' function with the appropriate
usat/ssat instruction if the min and max are right.  However, GCC's
heuristic only kicks in for 'int' and won't work on 'unsigned', so some
gymnastics are required for optimal code generation on ARMv7M.

Still, it seems better than the assembly versions.